% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mains.R
\name{BOP2FE_coprimary}
\alias{BOP2FE_coprimary}
\title{BOP2-FE design for co-primary endpoint}
\usage{
BOP2FE_coprimary(
  H0,
  H1,
  n,
  lambda = NULL,
  gamma = NULL,
  eta = NULL,
  method = "power",
  nsim = NULL,
  seed = NULL
)
}
\arguments{
\item{H0}{Response rate under the null (Response - PFS6, Response - no PFS6, No response - PFS6, No response - No PFS6)}

\item{H1}{Response rate under the alternative (Response - PFS6, Response - no PFS6, No response - PFS6, No response - No PFS6)}

\item{n}{A numeric vector representing the additional patients enrolled at each interim analysis. 
The value at index `i` indicates the number of new patients added at interim analysis `i`. 
The total sample size at interim `i` is the cumulative sum of the values in `n` up to that index. 
For example, for four interim analyses with total sample sizes of 10, 15, 20, and 30, 
the vector would be represented as `n = c(10, 5, 5, 10)`, where:
- 10 is the number of patients enrolled at interim 1,
- 5 (15 - 10) is the additional number of patients enrolled at interim 2,
- 5 (20 - 15) is the additional number of patients enrolled at interim 3,
- 10 (30 - 20) is the additional number of patients enrolled at interim 4.}

\item{lambda}{A numeric value for parameter `lambda` of the cut-off probability (i.e common for both efficacy and futility cut-off probability)}

\item{gamma}{A numeric value for parameter `gamma` of the cut-off probability for futility}

\item{eta}{A numeric value for parameter `eta` of the cut-off probability for efficacy}

\item{method}{A character string specifying the method to use for calculating cutoff values.
Options are "power" (default) or "OF" for "O'Brien-Fleming".}

\item{nsim}{number of simulation}

\item{seed}{for reproducibility}
}
\value{
A list of class \code{"BOP2FE"} containing the following elements:
\item{boundary}{A table of stopping boundaries.}
\item{Oc}{A table of operating characteristics.}
\item{plot}{A plot that shows the boundaries graphically along with tables of 
boundary values and operating characteristics}
}
\description{
Computes stopping boundaries and operating characteristics of Bayesian optimal phase II
design with efficacy and futility stopping for a co-primary endpoint given optimal
design parameter values that are identified to optimize power while controlling
type I error rate at a specified value. The optimal parameters (i.e., lambda, gamma, and eta)
can be obtained by calling `search_optimal_pars_coprimary()` function. The function produce a plot
that can be saved as a pdf file for documentation.
}
\examples{
\dontrun{
# Example with 7 interim looks
BOP2FE_coprimary(H0 = c(0.05,0.05,0.15,0.75), H1 =c(0.15,0.15,0.20,0.50), n=c(10,5,5,5,5,5,5), 
lambda = 0.95, gamma=1,seed = 123)
}
}
