% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mains.R
\name{BOP2FE_jointefftox}
\alias{BOP2FE_jointefftox}
\title{BOP2-FE design for joint efficacy and toxicity endpoint}
\usage{
BOP2FE_jointefftox(
  H0,
  n,
  lambda = NULL,
  gamma = NULL,
  eta = NULL,
  method = "power",
  nsim = NULL,
  seed = NULL
)
}
\arguments{
\item{H0}{Response rate under the null (toxicity - OR, no toxicity - OR, toxicity - no OR, no toxicity - No OR)}

\item{n}{vector of sample size at each interim look (Note that the sample size at interim i is the difference between sample size at interim i and at interim i-1)}

\item{lambda}{optimal value for lambda of the cut-off probability}

\item{gamma}{optimal value for gamma of the cut-off probability}

\item{eta}{optimal value for eta of the cut-off probability}

\item{method}{A character string specifying the method to use for calculating cutoff values.
Options are "power" (default) or "OF" for "O'Brien-Fleming".}

\item{nsim}{number of simulation}

\item{seed}{for reproducibility}
}
\value{
A list of class \code{"BOP2FE"} containing the following elements:
\item{boundary}{A table of stopping boundaries.}
\item{Oc}{A table of operating characteristics.}
\item{plot}{A plot that shows the boundaries graphically along with tables of 
boundary values and operating characteristics}
}
\description{
Computes stopping boundaries and operating characteristics of Bayesian optimal phase II
design with efficacy and futility stopping for a joint efficacy and toxicity endpoint given optimal
design parameter values that are identified to optimize power while controlling
type I error rate at a specified value. The optimal parameters (i.e., lambda, gamma, and eta)
can be obtained by calling `Optimal_pars_jointefftox()` function.
}
\examples{
\dontrun{
# Example with 7 interim looks
BOP2FE_jointefftox(H0 = c(0.15, 0.30, 0.15, 0.40), n=c(10,5,5,5,5,5,5), 
lambda = 0.7, gamma=1, seed = 123)
}
}
