% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mains.R
\name{BOP2FE_nested}
\alias{BOP2FE_nested}
\title{BOP2-FE design for nested (ordinal) endpoint}
\usage{
BOP2FE_nested(
  CR0,
  CRPR0,
  n,
  lambda = NULL,
  gamma = NULL,
  eta = NULL,
  method = NULL,
  nsim = NULL,
  seed = NULL
)
}
\arguments{
\item{CR0}{complete remission rate under the null (CR)}

\item{CRPR0}{complete remission or partial remission rate under the null (CR or PR)}

\item{n}{vector of sample size at each interim look (Note that the sample size at interim i is the difference between sample size at interim i and at interim i-1)}

\item{lambda}{optimal value for lambda of the cut-off probability}

\item{gamma}{optimal value for gamma of the cut-off probability}

\item{eta}{optimal value for eta of the cut-off probability}

\item{method}{type of function to be used for the cut off probability for superiority. The default is "OBrien-Fleming" function. method=power is an alternative}

\item{nsim}{number of simulation}

\item{seed}{for reproducibility}
}
\value{
A list of class \code{"BOP2FE"} containing the following elements:
\item{boundary}{A table of stopping boundaries.}
\item{Oc}{A table of operating characteristics.}
\item{plot}{A plot that shows the boundaries graphically along with tables of 
boundary values and operating characteristics}
}
\description{
#' Computes stopping boundaries and operating characteristics of Bayesian optimal phase II
design with efficacy and futility stopping for a nested (ordinal) endpoint given optimal
design parameter values that are identified to optimize power while controlling
type I error rate at a specified value. The optimal parameters (i.e., lambda, gamma, and eta)
can be obtained by calling `Optimal_pars_nested()` function.
}
