% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_coprimary.R
\name{boundary_coprimary}
\alias{boundary_coprimary}
\title{Boundary values for co primary Endpoint}
\usage{
boundary_coprimary(
  H0,
  a,
  n,
  nIA,
  lambda,
  gamma,
  eta = NULL,
  method = "power",
  seed = 123
)
}
\arguments{
\item{H0}{Response rate under the null (Response - PFS6, Response - no PFS6, No response - PFS6, No response - No PFS6)}

\item{a}{alpha values for the beta prior (i.e. usually set to the null response rate)}

\item{n}{sample size at each  interim analysis}

\item{nIA}{number of interim analysis}

\item{lambda}{optimal value for lambda of the cut-off probability}

\item{gamma}{optimal value for gamma of the cut-off probability}

\item{eta}{optimal value for eta of the cut-off probability}

\item{method}{type of function to be used for the cut off probability for superiority. The default is "power" type function. method=OF is an alternative for "O'Brien-Fleming"}

\item{seed}{seed number}
}
\description{
Boundary values for co primary Endpoint
}
