% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_posterior_binary}
\alias{calculate_posterior_binary}
\title{function title- under construction}
\usage{
calculate_posterior_binary(xdata, H0, H1, n_stage)
}
\arguments{
\item{xdata}{a given data set}

\item{H0}{null hypothesis}

\item{H1}{alternative hypothesis}

\item{n_stage}{the number of interm analysis}
}
\description{
function title- under construction
}
