% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_posterior_nested}
\alias{calculate_posterior_nested}
\title{Compute Posterior Probability of Observing a given Response Under the null for a nested outcome}
\usage{
calculate_posterior_nested(xdata, H0, H1, n_stage)
}
\arguments{
\item{xdata}{A data frame containing the data for which posterior probabilities are to be calculated.}

\item{H0}{A numeric vector representing the null hypothesis probabilities.}

\item{H1}{A numeric vector representing the alternative hypothesis probabilities.}

\item{n_stage}{An integer indicating the number of interim analysis in the trial.}
}
\value{
A data frame with additional columns for the posterior probabilities calculated for each stage.
}
\description{
This function computes the posterior probability of observing a given response under the null hypothesis
across multiple stages of a trial.
}
