% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_power_binary}
\alias{compute_power_binary}
\title{function title- under construction}
\usage{
compute_power_binary(H0, H1, n, nsim, n_stage)
}
\arguments{
\item{H0}{null hypothesis}

\item{H1}{alternative hypothesis}

\item{n}{sample size vector}

\item{nsim}{number of simulation}

\item{n_stage}{the number of interm analysis}
}
\description{
function title- under construction
}
