% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_power_jointefftox}
\alias{compute_power_jointefftox}
\title{Compute power and type I error rate for joint efficacy and toxicity endpoint}
\usage{
compute_power_jointefftox(
  H0,
  H1,
  n,
  nsim,
  lambda = NULL,
  gamma = NULL,
  eta = NULL,
  method = "power",
  seed = NULL
)
}
\arguments{
\item{H0}{the null hypothesis}

\item{H1}{the alternative hypothesis}

\item{n}{vector of sample sizes}

\item{nsim}{number of simulation to run}

\item{lambda}{lambda value for}

\item{gamma}{gamma value for}

\item{eta}{eta value for}

\item{method}{A character string specifying the method to use for calculating cutoff values.
Options are "power" (default) or "OF" for "O'Brien-Fleming".}

\item{seed}{for reproducability}
}
\description{
Compute power and type I error rate for joint efficacy and toxicity endpoint
}
