% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_power_nested}
\alias{compute_power_nested}
\title{function title - under construction}
\usage{
compute_power_nested(
  H0,
  H1,
  n,
  nsim,
  lambda = NULL,
  gamma = NULL,
  eta = NULL,
  method = "power",
  seed = NULL
)
}
\arguments{
\item{H0}{the null hypothesis}

\item{H1}{the alternative hypothesis}

\item{n}{vector of sample sizes}

\item{nsim}{number of simulation to run}

\item{lambda}{lambda value for}
}
\description{
function title - under construction
}
