% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oc_binary.R
\name{Oc_binary}
\alias{Oc_binary}
\title{Operating characteristics for for Binary Endpoint}
\usage{
Oc_binary(p, n, nsim, fb, sb, seed = 12345)
}
\arguments{
\item{p}{Response rate}

\item{n}{vector of sample size at each interim look (Note that the sample size at interim i is the difference between sample size at interim i and at interim i-1)}

\item{nsim}{number of simulation}

\item{fb}{vector of futility boundary at each interim}

\item{sb}{vector of superiority boundary at each interim}

\item{seed}{for responsibility}
}
\description{
Operating characteristics for for Binary Endpoint
}
