% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Optimal_pars_binary}
\alias{Optimal_pars_binary}
\title{Optimal parameters for binary outcome}
\usage{
Optimal_pars_binary(
  H0,
  H1,
  n,
  nsim,
  t1e,
  t2e,
  method = "power",
  lambda1,
  lambda2,
  grid1,
  gamma1,
  gamma2,
  grid2,
  eta1,
  eta2,
  grid3
)
}
\arguments{
\item{H0}{Probability of success under the null hypothesis.}

\item{H1}{Probability of success under the alternative hypothesis.}

\item{n}{A vector of sample sizes for each interim analysis.}

\item{nsim}{Number of simulation runs to perform.}

\item{t1e}{Desired Type - I error rate}

\item{t2e}{Desired Type - II error rate}

\item{method}{A character string specifying the method to use for calculating cutoff values.
Options are "power" (default) or "OF" for "O'Brien-Fleming".}

\item{lambda1}{starting value for lambda values to search}

\item{lambda2}{ending value for lambda values to search}

\item{grid1}{number of lambda values to consider between lambda1 and lambda2}

\item{gamma1}{starting value for gamma values to search}

\item{gamma2}{ending value for gamma values to search}

\item{grid2}{number of gamma values to consider between gamma1 and gamma2}

\item{eta1}{starting value for eta values to search}

\item{eta2}{ending value for eta values to search}

\item{grid3}{number of eta values to consider between eta1 and eta2}
}
\value{
A data frame
}
\description{
This function runs a grid search to find the optimal parameters for binary outcome that leads to 
the highest power while controlling the type - I error at the specified level.
}
