% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Optimal_pars_jointefftox}
\alias{Optimal_pars_jointefftox}
\title{Optimal parameters for joint efficacy and toxicity outcome}
\usage{
Optimal_pars_jointefftox(
  theta01,
  theta02,
  theta03,
  theta04,
  theta11,
  theta12,
  theta13,
  theta14,
  n,
  nsim,
  t1e,
  t2e,
  method,
  lambda1,
  lambda2,
  grid1,
  gamma1,
  gamma2,
  grid2,
  eta1,
  eta2,
  grid3
)
}
\arguments{
\item{theta01}{Probability of response and toxicity under the null hypothesis.}

\item{theta02}{Probability of response and no toxicity under the null hypothesis.}

\item{theta03}{Probability of no response and toxicity under the null hypothesis.}

\item{theta04}{Probability of no response and no toxicity under the null hypothesis.}

\item{theta11}{Probability of response and toxicity under the alternative hypothesis.}

\item{theta12}{Probability of response and no toxicity under the alternative hypothesis.}

\item{theta13}{Probability of no response and toxicity under the alternative hypothesis.}

\item{theta14}{Probability of no response and no toxicity under the alternative hypothesis.}

\item{n}{A vector of sample sizes for each interim analysis.}

\item{nsim}{Number of simulation runs to perform.}

\item{t1e}{Desired Type - I error rate}

\item{t2e}{Desired Type - II error rate}

\item{method}{method to use for cutoff}

\item{lambda1}{starting value for lambda values to search}

\item{lambda2}{ending value for lambda values to search}

\item{grid1}{number of lambda values to consider between lambda1 and lambda2}

\item{gamma1}{starting value for gamma values to search}

\item{gamma2}{ending value for gamma values to search}

\item{grid2}{number of gamma values to consider between gamma1 and gamma2}

\item{eta1}{starting value for eta values to search}

\item{eta2}{ending value for eta values to search}

\item{grid3}{number of eta values to consider between eta1 and eta2}
}
\value{
A data frame
}
\description{
This function runs a grid search to find the optimal parameters for joint efficacy and toxicity outcome that leads to the highest power while controlling the type - I error at the specified level.
}
