% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Optimal_pars_nested}
\alias{Optimal_pars_nested}
\title{Optimal parameters for nested outcome}
\usage{
Optimal_pars_nested(
  CR0,
  CRPR0,
  CR1,
  CRPR1,
  n,
  nsim,
  t1e,
  t2e,
  method,
  lambda1,
  lambda2,
  grid1,
  gamma1,
  gamma2,
  grid2,
  eta1,
  eta2,
  grid3
)
}
\arguments{
\item{CR0}{Probability of success for the first outcome under the null hypothesis.}

\item{CRPR0}{Probability of success for the first or second outcome under the null hypothesis.}

\item{CR1}{Probability of success for the first outcome under the alternative hypothesis.}

\item{CRPR1}{Probability of success for the first or second outcome under the alternative hypothesis.}

\item{n}{A vector of sample sizes for each interim analysis.}

\item{nsim}{Number of simulation runs to perform.}

\item{t1e}{Desired Type - I error rate}

\item{t2e}{Desired Type - II error rate}

\item{method}{method to use for cutoff}

\item{n_stage}{Number of interim analysis in the simulation.}

\item{lambda}{Range of lambda values to search}

\item{gamma}{Range of gamma values to search}

\item{eta}{Range of eta values to search if method is power}

\item{seed}{An integer seed for reproducibility of the random number generation.}
}
\value{
A data frame
}
\description{
This function runs a grid search to find the optimal parameters for nested or an ordinal outcome that leads to the highest power while controlling the type - I error at the specified level.
}
