% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mains.R
\name{plot.bop2fe}
\alias{plot.bop2fe}
\title{Plot the cut-off probability and simulation results for BOP2FE designs}
\usage{
\method{plot}{bop2fe}(x, ...)
}
\arguments{
\item{x}{the object returned by BOP2FE_xx}

\item{...}{additional parameters}
}
\value{
\code{plot()} returns a figure depending on the object entered
}
\description{
Plot the objects returned by other functions, including (1) cut-off probability;
(2) boundary values;
(3) operating characteristics
}
