% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_data_binary}
\alias{simulate_data_binary}
\title{Simulate binary endpoint data}
\usage{
simulate_data_binary(p, n_stage, n, nsim, seed = 23456)
}
\arguments{
\item{p}{Probability of success.}

\item{n_stage}{Number of interim analysis in the simulation.}

\item{n}{A vector of sample sizes for each interim analysis.}

\item{nsim}{Number of simulation runs to perform.}

\item{seed}{An integer seed for reproducibility of the random number generation.}
}
\value{
A data frame containing the simulated data for each observation, including cumulative sums for each stage.
}
\description{
This function simulates binary data across multiple stages based on specified probabilities and sample sizes.
}
