% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_data_coprimary}
\alias{simulate_data_coprimary}
\title{Simulate co-primary endpoint data}
\usage{
simulate_data_coprimary(p1, p2, p3, p4, n_stage, n, nsim, seed = 23456)
}
\arguments{
\item{p1}{Probability of response and progression-free survival at 6 months (PFS6).}

\item{p2}{Probability of response and no PFS6 .}

\item{p3}{Probability of no response and PFS6.}

\item{p4}{Probability of no response and no PFS6.}

\item{n_stage}{Number of interim analysis in the simulation.}

\item{n}{A vector of sample sizes for each interim analysis.}

\item{nsim}{Number of simulation runs to perform.}

\item{seed}{An integer seed for reproducibility of the random number generation.}
}
\value{
A data frame containing the simulated data for each observation, including cumulative sums for each stage.
}
\description{
This function simulates co-primary endpoint data across multiple stages based on specified probabilities and sample sizes.
}
\examples{
# Example usage of the simulate_data_nested function
result <- simulate_data_nested(p1 = 0.5, p2 = 0.3, p3 = 0.2,
 n_stage = 3, n = c(100, 150, 200), nsim = 10)
print(result)
}
