% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mains.R
\name{BOP2FE_binary}
\alias{BOP2FE_binary}
\title{BOP2 design for binary endpoint}
\usage{
BOP2FE_binary(
  H0,
  n,
  lambda = NULL,
  gamma = NULL,
  eta = NULL,
  method = NULL,
  nsim = NULL,
  seed = NULL
)
}
\arguments{
\item{H0}{Response rate}

\item{n}{vector of sample size at each interim look (Note that the sample size at interim i is the difference between sample size at interim i and at interim i-1)}

\item{lambda}{optimal value for lambda of the cut-off probability}

\item{gamma}{optimal value for gamma of the cut-off probability}

\item{eta}{optimal value for eta of the cut-off probability}

\item{method}{type of function to be used for the cut off probability for superiority. The default is power function}

\item{nsim}{number of simulation}

\item{seed}{for reproducibility}
}
\description{
BOP2 design for binary endpoint
}
