% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_binary.R
\name{boundary_binary}
\alias{boundary_binary}
\title{Boundary values for Binary Endpoint}
\usage{
boundary_binary(
  H0,
  a1,
  b1,
  nIA,
  n,
  lambda,
  gamma,
  eta = NULL,
  method = "power",
  seed = 123
)
}
\arguments{
\item{H0}{Response rate under the null}

\item{a1}{alpha values for the beta prior (i.e. usually set to the null response rate)}

\item{b1}{beta values for the beta prior (i.e, usually set to 1-null response rate)}

\item{nIA}{number of interim analysis}

\item{n}{sample size at each interim look}

\item{lambda}{optimal value for lambda of the cut-off probability}

\item{gamma}{optimal value for gamma of the cut-off probability}

\item{eta}{optimal value for eta of the cut-off probability}

\item{method}{type of function to be used for the cut off probability for superiority. The default is "power" type function. method=OF is an alternative for "O'Brien-Fleming"}

\item{seed}{for reproducibility}
}
\description{
Boundary values for Binary Endpoint
}
