% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_jointefftox.R
\name{boundary_jointefftox}
\alias{boundary_jointefftox}
\title{Boundary values for joint Endpoint}
\usage{
boundary_jointefftox(
  H0,
  a,
  n,
  nIA,
  lambda,
  gamma,
  eta,
  method = "power",
  seed = 123
)
}
\arguments{
\item{H0}{Response rate under the null (toxicity - OR, no toxicity - OR, toxicity - no OR, no toxicity - No OR)}

\item{a}{alpha values for the beta prior (i.e. usually set to the null response rate)}

\item{n}{sample size at each  interim analysis}

\item{nIA}{number of interim analysis}

\item{lambda}{optimal value for lambda of the cut-off probability}

\item{gamma}{optimal value for gamma of the cut-off probability}

\item{eta}{optimal value for eta of the cut-off probability}

\item{method}{method for the cut-off probability}

\item{seed}{seed number}
}
\description{
Boundary values for joint Endpoint
}
