% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_nested.R
\name{boundary_nested}
\alias{boundary_nested}
\title{Boundery values for Nested Endpoint}
\usage{
boundary_nested(H0, a, n, nIA, lambda, gamma, eta, method = NULL, seed = 123)
}
\arguments{
\item{H0}{Response rate under the null}

\item{a}{alpha values for the beta prior (i.e. usually set to the null response rate)}

\item{n}{sample size at each  interim analysis}

\item{nIA}{number of interim analysis}

\item{lambda}{optimal value for lambda of the cut-off probability}

\item{gamma}{optimal value for gamma of the cut-off probability}

\item{eta}{optimal value for eta of the cut-off probability}

\item{method}{method for the cut-off probability}

\item{seed}{seed number}
}
\description{
Boundery values for Nested Endpoint
}
