% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_posterior_binary}
\alias{calculate_posterior_binary}
\title{Compute Posterior Probability of Observing a given Response Under the null for a binary outcome}
\usage{
calculate_posterior_binary(xdata, H0, H1, n_stage, n)
}
\arguments{
\item{xdata}{a given data set}

\item{H0}{null hypothesis}

\item{H1}{alternative hypothesis}

\item{n_stage}{the number of interim analysis}

\item{n}{A vector of sample sizes for each interim analysis.}
}
\description{
Compute Posterior Probability of Observing a given Response Under the null for a binary outcome
}
