% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_posterior_coprimary}
\alias{calculate_posterior_coprimary}
\title{Compute Posterior Probability of Observing a given Response Under the null for a co-primary outcome}
\usage{
calculate_posterior_coprimary(xdata, H0, H1, n_stage)
}
\arguments{
\item{xdata}{A data frame containing the data for which posterior probabilities are to be calculated.}

\item{H0}{A numeric vector representing the null hypothesis probabilities.}

\item{H1}{A numeric vector representing the alternative hypothesis probabilities.}

\item{n_stage}{An integer indicating the number of interim analysis in the trial.}
}
\value{
A data frame with additional columns for the posterior probabilities calculated for each stage.
}
\description{
This function computes the posterior probability of observing a given response under the null hypothesis
across multiple stages of a trial.
}
\examples{
# Example usage of the calculate_posterior_coprimary function
#xdata <- data.frame(tsum11 = c(10, 20), tsum12 = c(5, 15), tsum13 = c(2, 8))
#h0 <- c(0.5, 0.3, 0.1,0.1)
#h1 <- c(0.6, 0.4, 0.1,0.1)
#result <- calculate_posterior_coprimary(xdata, h0, h1, n_stage = 3)
#print(result)
}
