% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oc_coprimary.R
\name{Oc_coprimary}
\alias{Oc_coprimary}
\title{Operating characteristics for for co primary Endpoint}
\usage{
Oc_coprimary(p1, p2, p3, p4, n_stage, n, nsim, fb, sb, seed = 12345)
}
\arguments{
\item{p1}{Response rate (Response - PFS6)}

\item{p2}{Response rate (Response - no PFS6)}

\item{p3}{Response rate (No response - PFS6)}

\item{p4}{Response rate (No response - no PFS6)}

\item{n_stage}{number of interim analysis}

\item{n}{vector of sample size at each interim look (Note that the sample size at interim i is the difference between sample size at interim i and at interim i-1)}

\item{nsim}{number of simulation}

\item{fb}{vector of futility boundary at each interim}

\item{sb}{vector of superiority boundary at each interim}

\item{seed}{for reproducibility}
}
\description{
Operating characteristics for for co primary Endpoint
}
