% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oc_nested.R
\name{Oc_nested}
\alias{Oc_nested}
\title{Operating characteristics for for nested Endpoint}
\usage{
Oc_nested(p1, p2, p3, n_stage, n, nsim, fb, sb, seed = 12345)
}
\arguments{
\item{p1}{Response rate (ORR)}

\item{p2}{Response rate (DCR - ORR)}

\item{p3}{Response rate (1-DCR)}

\item{n_stage}{number of interim analysis}

\item{n}{vector of sample size at each interim look (Note that the sample size at interim i is the difference between sample size at interim i and at interim i-1)}

\item{nsim}{number of simulation}

\item{fb}{vector of futility boundary at each interim}

\item{sb}{vector of superiority boundary at each interim}

\item{seed}{for reproducibility}
}
\description{
Operating characteristics for for nested Endpoint
}
