% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{search_optimal_pars_efftox}
\alias{search_optimal_pars_efftox}
\title{Search optimal parameters for joint efficacy and toxicity endpoint}
\usage{
search_optimal_pars_efftox(
  H0,
  H1,
  n,
  nsim,
  t1e = NULL,
  method = "power",
  lambda1,
  lambda2,
  grid1,
  gamma1,
  gamma2,
  grid2,
  eta1,
  eta2,
  grid3,
  seed = NULL
)
}
\arguments{
\item{H0}{A numeric vector representing the null response rates for different outcomes, specified in the following order:
- `H0[1]`: Response - toxicity,
- `H0[2]`: Response - no toxicity,
- `H0[3]`: No Response - toxicity,
- `H0[4]`: No Response - no toxicity}

\item{H1}{A numeric vector representing the null response rates for different outcomes, specified in the following order:
- `H1[1]`: Response - toxicity,
- `H1[2]`: Response - no toxicity,
- `H1[3]`: No Response - toxicity,
- `H1[4]`: No Response - no toxicity.}

\item{n}{A numeric vector representing the additional patients enrolled at each interim analysis. 
The value at index `i` indicates the number of new patients added at interim analysis `i`. 
The total sample size at interim `i` is the cumulative sum of the values in `n` up to that index. 
For example, for four interim analyses with total sample sizes of 10, 15, 20, and 30, 
the vector would be represented as `n = c(10, 5, 5, 10)`, where:
- 10 is the number of patients enrolled at interim 1,
- 5 (15 - 10) is the additional number of patients enrolled at interim 2,
- 5 (20 - 15) is the additional number of patients enrolled at interim 3,
- 10 (30 - 20) is the additional number of patients enrolled at interim 4.}

\item{nsim}{number of simulation. A value at least 1000 for better result}

\item{t1e}{Desired Type - I error rate. If specified it will only return results with type I error rate less the specified value}

\item{method}{A character string specifying the method to use for calculating cutoff values for the efficacy stoping.
Options are "power" (default) or "OF" for "O'Brien-Fleming".}

\item{lambda1}{starting value for `lambda` values to search}

\item{lambda2}{ending value for `lambda` values to search}

\item{grid1}{number of `lambda` values to consider between lambda1 and lambda2}

\item{gamma1}{starting value for `gamma` values to search}

\item{gamma2}{ending value for `gamma` values to search}

\item{grid2}{number of `gamma` values to consider between gamma1 and gamma2}

\item{eta1}{starting value for `eta` values to search}

\item{eta2}{ending value for `eta` values to search}

\item{grid3}{number of eta values to consider between eta1 and eta2}

\item{seed}{for reproducibility}
}
\value{
A data frame
}
\description{
`search_optimal_pars_efftox()` is a helper function to run a grid search to find the optimal parameters for futility and efficacy stopping cut-off probabilities. The function runs on 
a single core and can take a long time if the search space is too wide. If one has access to multiple cores, one can directly call
 `get_boundary_oc_efftox()` function and run it in parallel.
}
