% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_data_nested}
\alias{simulate_data_nested}
\title{Simulate nested endpoint data}
\usage{
simulate_data_nested(p1, p2, p3, n_stage, n, nsim, seed = 23456)
}
\arguments{
\item{p1}{Probability of success for the first outcome.}

\item{p2}{Probability of success for the second outcome.}

\item{p3}{Probability of success for the third outcome.}

\item{n_stage}{Number of interim analysis in the simulation.}

\item{n}{A vector of sample sizes for each interim analysis.}

\item{nsim}{Number of simulation runs to perform.}

\item{seed}{An integer seed for reproducibility of the random number generation.}
}
\value{
A data frame containing the simulated data for each observation, including cumulative sums for each stage.
}
\description{
This function simulates nested data across multiple stages based on specified probabilities and sample sizes.
}
